Moving the network selector into a seperate component
